Updated module augmentation for typing package.
