+++ packages/ckeditor5-typing/src/augmentation.ts
@@ -1,0 +1,39 @@
+/**
+ * @license Copyright (c) 2003-2023, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+ */
+
+import type Delete from './delete';
+import type DeleteCommand from './deletecommand';
+import type Input from './input';
+import type InsertTextCommand from './inserttextcommand';
+import type TextTransformation from './texttransformation';
+import type TwoStepCaretMovement from './twostepcaretmovement';
+import type Typing from './typing';
+import type { TypingConfig } from './typingconfig';
+
+declare module '@ckeditor/ckeditor5-core' {
+	interface EditorConfig {
+
+		/**
+		 * The configuration of the typing features. Used by the features from the `@ckeditor/ckeditor5-typing` package.
+		 *
+		 * Read more in {@link module:typing/typingconfig~TypingConfig}.
+		 */
+		typing?: TypingConfig;
+	}
+
+	interface CommandsMap {
+		deleteForward: DeleteCommand;
+		delete: DeleteCommand;
+		insertText: InsertTextCommand;
+	}
+
+	interface PluginsMap {
+		[ Delete.pluginName ]: Delete;
+		[ Input.pluginName ]: Input;
+		[ TextTransformation.pluginName ]: TextTransformation;
+		[ TwoStepCaretMovement.pluginName ]: TwoStepCaretMovement;
+		[ Typing.pluginName ]: Typing;
+	}
+}
