+++ packages/ckeditor5-typing/src/twostepcaretmovement.ts
@@ -499,9 +499,3 @@ function isBetweenDifferentAttributes( position: Position, attributes: Set<strin
 	}
 	return false;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ TwoStepCaretMovement.pluginName ]: TwoStepCaretMovement;
-	}
-}
