+++ packages/ckeditor5-typing/src/typing.ts
@@ -29,9 +29,3 @@ export default class Typing extends Plugin {
 		return 'Typing';
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ Typing.pluginName ]: Typing;
-	}
-}
